Adjusting controls in HED Filter
Changed the controls in tab filter to allow for the change of side outputs results of each filter.
Before this change, there was an error in apply filter.
The error:
the apply_filter method expects the first parameter to be an image (or None), not an integer. This mismatch causes image to be an integer, leading to the error when the method checks image.shape.

To prevent the currentIndexChanged signal from passing the integer argument to apply_filter,

I connected it it through an intermediate method.