Adding Control Points instead of thousands of Nodes
To avoid nodes in all lines. The goal is to make the lines as vector graphics as much as possible for this reason put the nodes as defined by the users instead of adding it automatically.

So removing nodes as a whole, and adding an option to right click on line to add nodes or control points.